rename class and interface
